device: Allow inner container to have a different device path from the host
Currently, updateSpecDevice() assumes that the proper device path in the
inner container is the same as the device path specified in the outer OCI
spec on the host.

Usually that's correct.  However for VFIO group devices we actually need
the container to see the VM's device path, since it's normal to correlate
that with IOMMU group information from sysfs which will be different in the
guest and which we can't namespace away.

So, add an extra "finalPath" parameter to updateSpecDevice to allow callers
to chose the device path that should be used for the inner container.  All
current callers pass the same thing as ContainerPath, but that will change
in future.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>