device: Rebind VFIO devices to VFIO driver inside guest
VFIO devices can be added to a Kata container and they will be passed
through to the sandbox guest.  However, inside the guest those devices
will generally bind to a native guest driver, so they will no longer
appear as VFIO devices within the guest.  This behaviour differs from
runc or other conventional container runtimes.

This code allows the agent to match the behaviour of other runtimes,
if instructed to by kata-runtime.  VFIO devices its informed about
with the "vfio" type instead of the existing "vfio-vm" type will be
rebound to the vfio-pci driver within the guest.

Fixes: #833

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>